pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "traa"
version = "0.1.0"
description = "A Python library for capturing and interacting with the desktop screen"
readme = "README.md"
keywords = [
    "Screen Capturer", "Snapshot", "Desktop Capturer"
]
license = { text = "BSD 3-Clause License"}
authors = [
  { name = "Sylar", email = "peilinok@gmail.com" },
]
requires-python = ">=3.10"
dependencies = [
    "cffi>=1.17.1",
    "numpy>=2.2.3",
    "pillow>=11.1.0",
]

[project.urls]
Homepage = "https://github.com/opentraa/traa-py"
Issues = "https://github.com/opentraa/traa-py/issues"

[tool.mypy]
python_version = "3.9"
pretty = true
disallow_any_generics = true
enable_error_code = "ignore-without-code"
extra_checks = true
follow_imports = "silent"
warn_redundant_casts = true
warn_unreachable = true
warn_unused_ignores = true